chore: moved PromptGenerator to prompts
